# docs/swagger.yaml
basePath: /
definitions:
  controllers.Login:
    properties:
      device_id:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
  controllers.Logout:
    properties:
      device_id:
        type: string
    type: object
  controllers.Refresh:
    properties:
      refresh_token:
        type: string
    type: object
...
  models.CreateUser:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  models.LoginResponse:
    properties:
      access_token:
        type: string
      device_id:
        type: string
      expiry_date:
        type: integer
      refresh_token:
        type: string
    type: object
  models.Permission:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      email:
        type: string
      enabled:
        type: boolean
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      provider:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      updated_at:
        type: string
      updated_by:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    name: API Support
  description: This is an RBAC based full fledge API serice for managing users, roles
    and permissions in the system
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Truvest Identity Management Service APIs
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Login to the system. It returns the accessToken, refreshToken,
        expiry and device_id in a JSON format. You need to pass email and password
        along with a unique device_id. You can use device fingerprint or browser fingerprint
        to generate a unique device_id. The access and refresh token will be mapped
        to this id so that whenever the user logs out, it can detect the device from
        which the user is logging out in case of multi-session login and can expire
        the JWT for that specific device.
      parameters:
      - description: Login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/controllers.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
      summary: Login to the system
      tags:
      - Login
  /logout:
    post:
      consumes:
      - application/json
      description: User can logout from the system using this API. User need to pass
        device_id so that he can be logged out of the specific device during multi-session
        login. This will expire the JWT for the device from where he has logged in.
      parameters:
      - description: Logout
        in: body
        name: logout
        required: true
        schema:
          $ref: '#/definitions/controllers.Logout'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Logout from the system
      tags:
      - Login
  /logout/{provider}:
    post:
      consumes:
      - application/json
      description: User can logout from the system and OAuth provider using this API.
        User need to pass the provider as param and last device_id with which user
        logged in.
      parameters:
      - description: OAuth provider
        in: path
        name: provider
        required: true
        type: string
      - description: Logout
        in: body
        name: logout
        required: true
        schema:
          $ref: '#/definitions/controllers.Logout'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Logout from the system using OAuth provider
      tags:
      - OAuth
...
  /refresh:
    post:
      consumes:
      - application/json
      description: Refresh JWT token to Login to the system. It returns the accessToken,
        refreshToken and expiry in a JSON format. You need to pass the refresh token
        from the previous call.
      parameters:
      - description: Refresh
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/controllers.Refresh'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
      security:
      - ApiKeyAuth: []
      summary: Refresh JWT Token to Login to the system
      tags:
      - Login
...
  /signup:
    post:
      consumes:
      - application/json
      description: Signup as a user in the system. If Password is not passed while
        creating User, then it creates User with auto-generated password. The Password
        is then sent through Email.
      parameters:
      - description: Create User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.UserResponse'
      summary: Signup as a user in the system
      tags:
      - SignUp
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"